Passport.js removal and JS reorganization
reorganized some of the js to reflect mistah's reorganization of the html

and a few needed backEnd tweaks as well, such as,

fixed a name collision problem with the "username" of the logged in user, and the "username" of the user whose page you are looking at. Now it's "authorName".

in tracking a slight routing error, I ended up down a rabbit hole and noticed that passport.js was doing some weird stuff I didn't want it to so I removed it and am now just using node's built in sessions thing and bcrypt for hashing and cut out the middleman(ware) of Passport.

oh and a placeholder favicon so it'll stop bothering me about not finding a favicon at the default favicon route